--- notification/src/main/java/org/bookwoori/notification/config/AsyncConfig.java
@@ -1,24 +1,0 @@
-package org.bookwoori.notification.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.annotation.AsyncConfigurerSupport;
-import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-
-import java.util.concurrent.Executor;
-
-@Configuration
-@EnableAsync
-public class AsyncConfig extends AsyncConfigurerSupport {
-
-    @Override
-    public Executor getAsyncExecutor() {
-        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-        executor.setCorePoolSize(10);
-        executor.setMaxPoolSize(30);
-        executor.setQueueCapacity(50);
-        executor.setThreadNamePrefix("NOTIFICATION-SERVER-SYNC-");
-        executor.initialize();
-        return executor;
-    }
-}
