--- notification/src/main/java/org/bookwoori/notification/config/MongoConfig.java
@@ -1,25 +1,0 @@
-package org.bookwoori.notification.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.mongodb.MongoDatabaseFactory;
-import org.springframework.data.mongodb.core.MongoTemplate;
-import org.springframework.data.mongodb.core.SimpleMongoClientDatabaseFactory;
-
-@Configuration
-public class MongoConfig {
-
-    @Value("${spring.data.mongodb.uri}")
-    private String connection;
-
-    @Bean
-    public MongoDatabaseFactory mongoDatabaseFactory() {
-        return new SimpleMongoClientDatabaseFactory(connection);
-    }
-
-    @Bean
-    public MongoTemplate mongoTemplate() {
-        return new MongoTemplate(mongoDatabaseFactory());
-    }
-}
