+++ notification/src/main/java/org/bookwoori/notification/domain/dto/request/BatchChannelMessageRequestDto.java
@@ -1,0 +1,12 @@
+package org.bookwoori.notification.domain.dto.request;
+
+import java.time.LocalDateTime;
+
+
+public record BatchChannelMessageRequestDto(
+        LocalDateTime time,
+        ChannelMessageRequestDto request
+
+) {
+
+}
