+++ notification/src/main/java/org/bookwoori/notification/domain/service/ResponseService.java
@@ -1,0 +1,42 @@
+package org.bookwoori.notification.domain.service;
+
+
+import org.bookwoori.notification.domain.dto.response.CommonResponseDto;
+import org.bookwoori.notification.domain.dto.response.DataResponseDto;
+import org.bookwoori.notification.global.exception.CustomExceptionStatus;
+import org.springframework.stereotype.Service;
+
+@Service
+public class ResponseService {
+
+    public CommonResponseDto getSuccessResponse() {
+
+        return CommonResponseDto.builder()
+                .isSuccess(true)
+                .code(5000)
+                .message("요청에 성공하였습니다.")
+                .build();
+
+    }
+
+    public <T> DataResponseDto<Object> getDataResponse(T data) {
+
+        return DataResponseDto.builder()
+                .result(data)
+                .isSuccess(true)
+                .code(5000)
+                .message("요청에 성공하였습니다.")
+                .build();
+
+    }
+
+    public CommonResponseDto getExceptionResponse(CustomExceptionStatus status) {
+
+        return CommonResponseDto.builder()
+                .isSuccess(status.isSuccess())
+                .code(status.getCode())
+                .message(status.getMessage())
+                .build();
+
+    }
+}
