--- notification/src/main/java/org/bookwoori/notification/dto/request/EmojiMessageRequestDto.java
@@ -1,27 +1,0 @@
-package org.bookwoori.notification.dto.request;
-
-import jakarta.validation.constraints.NotNull;
-
-
-public record EmojiMessageRequestDto(
-        @NotNull
-        Long memberId,
-
-        @NotNull
-        String nickname,
-
-        @NotNull
-        String type,
-
-        @NotNull
-        String content,
-
-        @NotNull
-        Long reviewId,
-
-        @NotNull
-        String target
-) {
-
-
-}
