--- notification/src/main/java/org/bookwoori/notification/dto/response/CommonResponseDto.java
@@ -1,14 +1,0 @@
-package org.bookwoori.notification.dto.response;
-
-import lombok.Builder;
-
-@Builder
-public record CommonResponseDto(
-        Boolean isSuccess,
-        int code,
-        String message
-) {
-
-
-}
-
