--- notification/src/main/java/org/bookwoori/notification/exception/CustomException.java
@@ -1,10 +1,0 @@
-package org.bookwoori.notification.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class CustomException extends RuntimeException {
-    CustomExceptionStatus customExceptionStatus;
-}
