+++ notification/src/main/java/org/bookwoori/notification/global/exception/CustomException.java
@@ -1,0 +1,10 @@
+package org.bookwoori.notification.global.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class CustomException extends RuntimeException {
+    CustomExceptionStatus customExceptionStatus;
+}
