+++ notification/src/main/java/org/bookwoori/notification/global/util/DeviceType.java
@@ -1,0 +1,6 @@
+package org.bookwoori.notification.global.util;
+
+public class DeviceType {
+    public static final String WEB = "web";
+    public static final String IOS = "ios";
+}
