+++ notification/src/main/java/org/bookwoori/notification/global/util/MessageType.java
@@ -1,0 +1,6 @@
+package org.bookwoori.notification.global.util;
+
+public class MessageType {
+    public static final String TALK = "talk";
+    public static final String FILE = "file";
+}
