--- notification/src/main/java/org/bookwoori/notification/util/DeviceType.java
@@ -1,6 +1,0 @@
-package org.bookwoori.notification.util;
-
-public class DeviceType {
-    public static final String WEB = "web";
-    public static final String IOS = "ios";
-}
