--- notification/src/main/java/org/bookwoori/notification/util/MessageType.java
@@ -1,6 +1,0 @@
-package org.bookwoori.notification.util;
-
-public class MessageType {
-    public static final String TALK = "talk";
-    public static final String FILE = "file";
-}
